Renamed package to common
