+++ core/src/main/java/com/sap/ai/sdk/core/common/ClientError.java
@@ -1,4 +1,4 @@
-package com.sap.ai.sdk.core.commons;
+package com.sap.ai.sdk.core.common;
 
 import com.google.common.annotations.Beta;
 import javax.annotation.Nullable;
