+++ foundation-models/openai/src/main/java/com/sap/ai/sdk/foundationmodels/openai/OpenAiClientException.java
@@ -1,6 +1,6 @@
 package com.sap.ai.sdk.foundationmodels.openai;
 
-import com.sap.ai.sdk.core.commons.ClientException;
+import com.sap.ai.sdk.core.common.ClientException;
 import lombok.experimental.StandardException;
 
 /** Generic exception for errors occurring when using OpenAI foundation models. */
