+++ foundation-models/openai/src/main/java/com/sap/ai/sdk/foundationmodels/openai/model/OpenAiChatCompletionDelta.java
@@ -2,7 +2,7 @@ package com.sap.ai.sdk.foundationmodels.openai.model;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 import com.google.common.annotations.Beta;
-import com.sap.ai.sdk.core.commons.StreamedDelta;
+import com.sap.ai.sdk.core.common.StreamedDelta;
 import java.util.List;
 import javax.annotation.Nonnull;
 import javax.annotation.Nullable;
