+++ orchestration/src/main/java/com/sap/ai/sdk/orchestration/OrchestrationClient.java
@@ -9,9 +9,9 @@ import com.fasterxml.jackson.databind.node.ObjectNode;
 import com.google.common.annotations.Beta;
 import com.sap.ai.sdk.core.AiCoreService;
 import com.sap.ai.sdk.core.DeploymentResolutionException;
-import com.sap.ai.sdk.core.commons.ClientResponseHandler;
-import com.sap.ai.sdk.core.commons.ClientStreamingHandler;
-import com.sap.ai.sdk.core.commons.StreamedDelta;
+import com.sap.ai.sdk.core.common.ClientResponseHandler;
+import com.sap.ai.sdk.core.common.ClientStreamingHandler;
+import com.sap.ai.sdk.core.common.StreamedDelta;
 import com.sap.ai.sdk.orchestration.model.CompletionPostRequest;
 import com.sap.ai.sdk.orchestration.model.CompletionPostResponse;
 import com.sap.ai.sdk.orchestration.model.LLMModuleResult;
