+++ orchestration/src/main/java/com/sap/ai/sdk/orchestration/OrchestrationClientException.java
@@ -1,6 +1,6 @@
 package com.sap.ai.sdk.orchestration;
 
-import com.sap.ai.sdk.core.commons.ClientException;
+import com.sap.ai.sdk.core.common.ClientException;
 import lombok.experimental.StandardException;
 
 /** Exception thrown by the {@link OrchestrationClient} in case of an error. */
